Fix rcg_optl_omp declaration

+++ src/rcgpar.cpp
@@ -56,7 +56,7 @@ void check_input(const seamat::Matrix<double> &logl, const std::vector<double> &
     }
 }
 
-seamat::DenseMatrix<double> rcg_optl_omp(const seamat::Matrix<double> &logl, const std::vector<double> &log_times_observed, const std::vector<double> &alpha0, const double &tol, uint16_t max_iters, std::ostream &log) {
+seamat::DenseMatrix<double> rcg_optl_omp(const seamat::Matrix<double> &logl, const std::vector<double> &log_times_observed, const std::vector<double> &alpha0, const double &tol, size_t max_iters, std::ostream &log) {
     // Validate input data
     check_input(logl, log_times_observed, alpha0, tol, max_iters);
